pyproject.toml
[project]
name = "coverse"
version = "1.0.0"
description = "coverse"
authors = [{ name = "Eric Zhang" }]
readme = "README.md"
requires-python = "==3.11.*"
license = "MIT"

dependencies = [
    "torch==2.6.0",
    "transformers",
    "accelerate",
    "numpy>=1.27",
    "pandas>=2.2",
    "gradio",
    "openai",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "pandas-stubs>=2.3.2.250926",
    "tqdm>=4.67.1",
]

[project.scripts]
coverse = "coverse.cli.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["coverse*"]
exclude = ["config*"]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.ruff]
line-length = 100

